==> components/PS_SOTA/1a - Predictions using SOTA (PG + BA).ipynb <==
# Peptide generation + binding prediction using MHCflurry

This is essentially AP predictor from MHCflurry

1. The training data for PS is S6 dataset
2. The evaluation is done on Multiallelic-recent S1 data

import pandas as pd
import numpy as np
from mhcflurry import Class1PresentationPredictor
from tqdm import tqdm

# read training data
training_df = pd.read_csv('./../../data/PS/PS_df.csv', index_col=0)
training_df

# allele list
allele_ls = training_df['allele'].unique()
allele_ls

### load
predictor = Class1PresentationPredictor.load()
predictor.weights_dataframe


### Predict and make df
predictor_result_df = [None]*len(allele_ls)

for idx, allele in tqdm(enumerate(allele_ls)):
    allele_df = training_df[training_df['allele']==allele]
    
    predictor_result_df[idx] = predictor.predict(
                                    peptides=allele_df['peptide'].to_list(),
                                    alleles=[allele],
                                    verbose=0, throw=False)
    
predictor_result_df = pd.concat(predictor_result_df)
predictor_result_df = predictor_result_df.rename(columns={'best_allele':'allele'}) 
predictor_result_df = predictor_result_df.merge(training_df, on=['peptide','allele'])
predictor_result_df

# # save final df
# predictor_result_df = predictor_result_df.sort_values(['peptide','affinity'])
# predictor_result_df = predictor_result_df.drop_duplicates(['peptide','PS_y'])
# predictor_result_df.to_csv('./results/PS_training_MHCflurry_result.csv')

# check peptide count
predictor_result_df = pd.read_csv('./results/PS_training_MHCflurry_result.csv',index_col=0)
len(predictor_result_df['peptide'].unique()) == len(training_df['peptide'].unique())
