==> components/PS_SOTA/1c - Combined training dataset.ipynb <==
# Combining prediction from 3 SOTA + literature for training PS model

This is essentially AP prediction + BA prediction from MHCflurry + TAP from DeepTAP + Bashirova for TD

1. The training data for PS is S6 dataset

import pandas as pd
import numpy as np
from tqdm import tqdm

# read PG + BA data
mhcflurry_df = pd.read_csv('./results/PS_training_MHCflurry_result.csv',index_col=0)
mhcflurry_df['BA_proba'] = [1-np.log(affinity)/np.log(50000) for affinity in mhcflurry_df['affinity']]
mhcflurry_df['BA_y'] = [1 if affinity<500 else 0 for affinity in mhcflurry_df['affinity']]

mhcflurry_df = mhcflurry_df.rename(columns={'processing_score':'PG_proba', 'presentation_score':'PS_proba'})
mhcflurry_df['PG_y'] = [1 if proba>0.5 else 0 for proba in mhcflurry_df['PG_proba']]

# read TAP data
deeptap_df = pd.read_csv('./results/PS_peptides_DeepTAP_cla_predresult.csv')
deeptap_df = deeptap_df.rename(columns={'pred_label':'TAP_y', 'pred_score':'TAP_proba'})

# read TD data
bashirova_df = pd.read_csv('./../../data/TD/processed_data//TD_full.csv',index_col=0)
bashirova_df = bashirova_df.rename(columns={'HLA_full':'allele'})
bashirova_df['TD_y'] = [1 if i>2 else 0 for i in bashirova_df['MFI_ratio']]


# combine datasets
PS_training_df = mhcflurry_df.merge(deeptap_df, on='peptide').merge(bashirova_df, on='allele')
PS_training_df.to_csv('./results/PS_training.csv')
PS_training_df
