==> components/PS_SOTA/2b - Train LR model for PS using SOTA proba concepts.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm import tqdm
import pickle

from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split, StratifiedKFold
from sklearn.metrics import auc, roc_curve, roc_auc_score, f1_score, precision_recall_curve, accuracy_score, recall_score,  precision_score
from sklearn import preprocessing

# read training data
PS_training_df = pd.read_csv('./results/PS_training.csv', index_col=0)

# TD proba
min_max_scaler = preprocessing.MinMaxScaler()
PS_training_df['TD_proba'] = min_max_scaler.fit_transform(PS_training_df[['MFI_ratio']]).reshape(-1)
PS_training_df

# keep only relevant
PS_training_df = PS_training_df[['peptide',
                                 'allele',
                                 'PG_proba',
                                 'TAP_proba',
                                 'BA_proba',
                                 'TD_proba',
                                 'PG_y',
                                 'TAP_y',
                                 'BA_y',
                                 'TD_y',
                                 'PS_proba',
                                 'PS_y']]


##### NOTE Here PS_y is hit (real y) 

# Create X and y

X_y_df = PS_training_df[['PG_proba',
                            'TAP_proba',
                            'BA_proba',
                            'TD_proba',
                            'PS_y']].dropna()

X = X_y_df[['PG_proba',
            'TAP_proba',
            'BA_proba',
            'TD_proba']].to_numpy()

y = X_y_df['PS_y'].to_numpy()

# Create 
nfold = 10
skf = StratifiedKFold(n_splits=nfold, )

auroc_ls = []
auprc_ls = []
f1_ls = []
precision_ls = []
recall_ls = []

fold_df = {'auroc':[], 'auprc':[], 
           'f1':[], 
           'accuracy': [],
           'precision':[],  
           'recall':[]}

for i, (train_index, test_index) in enumerate(skf.split(X, y)):
    clf = LogisticRegression(random_state=0,fit_intercept=True)
    
    clf.fit(X[train_index], y[train_index])
    y_pred = clf.predict(X[test_index])
    y_proba_pred = clf.predict_proba(X[test_index])
    

    AUROC = roc_auc_score(y[test_index], y_proba_pred[:,1])
    fold_df['auroc'] = fold_df['auroc'] + [AUROC]
    
    # calculate precision-recall curve
    precision, recall, thresholds = precision_recall_curve(y[test_index], y_proba_pred[:,1])
    AUPRC = auc(recall, precision)
    fold_df['auprc'] = fold_df['auprc'] + [AUPRC]
    
    # calculate f1 score
    f1 = f1_score(y[test_index], y_pred)
    fold_df['f1'] = fold_df['f1'] + [f1]
    
    # calculate accuracy score
    accuracy = accuracy_score(y[test_index], y_pred)
    fold_df['accuracy'] = fold_df['accuracy'] + [accuracy]
    
    # calculate precision score
    precision = precision_score(y[test_index], y_pred)
    fold_df['precision'] = fold_df['precision'] + [precision]
    
    # calculate recall score
    recall = recall_score(y[test_index], y_pred)
    fold_df['recall'] = fold_df['recall'] + [recall]
    
    
fold_df = pd.DataFrame(fold_df)
fold_df['fold'] = [i for i in range(nfold)]
fold_df

# final model
clf = LogisticRegression(random_state=0,
                         fit_intercept=True,)
    
clf.fit(X, y)

# save model
with open('./models/PS_SOTA_clf_proba.pkl', 'wb') as f:
    pickle.dump(clf, f)
f.close()

# check if model is saved correctly
with open('./models/PS_SOTA_clf_proba.pkl', 'rb') as f:
    clf = pickle.load(f)
f.close()

clf
