==> components/PS_SOTA/3 - Predictions using SOTA (PT) benchmark.ipynb <==
# Peptide Transport using DeepTAP

The evaluation is done on Multiallelic-recent S1 data

import pandas as pd
import numpy as np

from tqdm import tqdm

# read training data
training_df = pd.read_csv('./../../data/BA/Data_S1.csv')
training_df = training_df[training_df['sample_group']=='MULTIALLELIC-RECENT']
training_df

# just keep the peptides
training_df[['peptide']].iloc[:200000].drop_duplicates().to_csv('./../../data/PS/PS_benchmark_peptides1.csv', index=None)
training_df[['peptide']].iloc[200000:400000].drop_duplicates().to_csv('./../../data/PS/PS_benchmark_peptides2.csv', index=None)
training_df[['peptide']].iloc[400000:600000].drop_duplicates().to_csv('./../../data/PS/PS_benchmark_peptides3.csv', index=None)
training_df[['peptide']].iloc[600000:800000].drop_duplicates().to_csv('./../../data/PS/PS_benchmark_peptides4.csv', index=None)
training_df[['peptide']].iloc[800000:1000000].drop_duplicates().to_csv('./../../data/PS/PS_benchmark_peptides5.csv', index=None)
training_df[['peptide']].iloc[1000000:1200000].drop_duplicates().to_csv('./../../data/PS/PS_benchmark_peptides6.csv', index=None)
training_df[['peptide']].iloc[1200000:1400000].drop_duplicates().to_csv('./../../data/PS/PS_benchmark_peptides7.csv', index=None)
training_df[['peptide']].iloc[1400000:1600000].drop_duplicates().to_csv('./../../data/PS/PS_benchmark_peptides8.csv', index=None)
training_df[['peptide']].iloc[1600000:1800000].drop_duplicates().to_csv('./../../data/PS/PS_benchmark_peptides9.csv', index=None)
training_df[['peptide']].iloc[1800000:2000000].drop_duplicates().to_csv('./../../data/PS/PS_benchmark_peptides10.csv', index=None)
training_df[['peptide']].iloc[2000000:2200000].drop_duplicates().to_csv('./../../data/PS/PS_benchmark_peptides11.csv', index=None)
training_df[['peptide']].iloc[2200000:].drop_duplicates().to_csv('./../../data/PS/PS_benchmark_peptides12.csv', index=None)

# diff = 100000
# peptide_df = training_df[['peptide']].drop_duplicates()

# for idx, stop in enumerate(range(0,len(peptide_df), diff)):
#     print(idx, stop, stop+diff)
#     peptide_df[['peptide']].iloc[stop:stop+diff].to_csv('./../../data/PS/PS_benchmark_peptides'+str(idx)+'.csv', index=None)

In terminal 


1. source activate deeptap310
2. cd /DeepTAP-main/
3. python deeptap.py -t cla -f input file -o output directory

input file - /exports/csce/eddie/inf/groups/ajitha_project/piyush/MHCCBM/MHCCBM/data/PS/PS_benchmark_peptides+idx.csv

output directory - /exports/csce/eddie/inf/groups/ajitha_project/piyush/MHCCBM/MHCCBM/components/PS_SOTA/results/

# # collate results
# prediction_df = [None]*24
# for idx, stop in enumerate(range(0,len(peptide_df), diff)):
#     prediction_df[idx] = pd.read_csv('./results//PS_benchmark_peptides'+str(idx)+'_DeepTAP_cla_predresult.csv')
    
# prediction_df = pd.concat(prediction_df)
# prediction_df

# save final result
prediction_df.to_csv('./results/PS_benchmark_peptides_DeepTAP_cla_predresult.csv')

prediction_df = pd.read_csv('./results/PS_benchmark_peptides_DeepTAP_cla_predresult.csv', index_col=0)
prediction_df
