==> components/PS_SOTA/4a - Performance on Benchmark dataset binary concept.ipynb <==
Benchmark for PS predictor is multiallelic-recent benchmark in Data S1

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm import tqdm
import pickle

from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split, StratifiedKFold
from sklearn.metrics import auc, roc_curve, roc_auc_score, f1_score, precision_recall_curve, accuracy_score, recall_score,  precision_score


# # read PG + BA data
# mhcflurry_df = pd.read_csv('./../../data/BA/Data_S1.csv',index_col=0)
# mhcflurry_df = mhcflurry_df[mhcflurry_df['sample_group']=='MULTIALLELIC-RECENT']
# mhcflurry_df['BA_proba'] = [1-np.log(affinity)/np.log(50000) for affinity in mhcflurry_df['mhcflurry2.ba']]
# mhcflurry_df['BA_y'] = [1 if affinity<500 else 0 for affinity in mhcflurry_df['mhcflurry2.ba']]

# mhcflurry_df = mhcflurry_df.rename(columns={'mhcflurry2.ap.without_flanks':'PG_proba', 
#                                             'mhcflurry2.ps.without_flanks':'PS_proba',
#                                             'mhcflurry2.ba_best_allele':'allele'})
# mhcflurry_df['PG_y'] = [1 if proba>0.5 else 0 for proba in mhcflurry_df['PG_proba']]
# mhcflurry_df['PS_y'] = [1 if proba>0.5 else 0 for proba in mhcflurry_df['PS_proba']]
# mhcflurry_df = mhcflurry_df[['peptide','allele','hit',
#                              'BA_proba','BA_y',
#                              'PG_proba','PG_y',
#                              'PS_proba','PS_y']]


# # read TAP data
# deeptap_df = pd.read_csv('./results/PS_benchmark_peptides_DeepTAP_cla_predresult.csv', index_col=0)
# deeptap_df = deeptap_df.rename(columns={'pred_label':'TAP_y', 'pred_score':'TAP_proba'})

# # read TD data
# bashirova_df = pd.read_csv('./../../data/TD/processed_data//TD_full.csv',index_col=0)
# bashirova_df = bashirova_df.rename(columns={'HLA_full':'allele'})
# bashirova_df['TD_y'] = [1 if i>2 else 0 for i in bashirova_df['MFI_ratio']]



# PS_benchmark_df = mhcflurry_df.merge(deeptap_df, on='peptide').merge(bashirova_df, on='allele')
# PS_benchmark_df.to_csv('./results/PS_benchmark.csv')
# PS_benchmark_df

# read training data
PS_benchmark_df = pd.read_csv('./results/PS_benchmark.csv', index_col=0)

# keep only relevant
PS_benchmark_df = PS_benchmark_df[['peptide',
                                 'allele',
                                 'PG_proba',
                                 'TAP_proba',
                                 'BA_proba',
                                 'PG_y',
                                 'TAP_y',
                                 'BA_y',
                                 'TD_y',
                                 'PS_proba',
                                 'PS_y',
                                 'hit']]

# Create X and y
X = PS_benchmark_df[['PG_y',
                    'TAP_y',
                    'BA_y',
                    'TD_y']].to_numpy()

y = PS_benchmark_df[['hit']].to_numpy()

# load model
with open('./models/PS_SOTA_clf_binary.pkl', 'rb') as f:
    clf = pickle.load(f)
f.close()

clf

clf.coef_, clf.intercept_

y_pred = clf.predict(X)
y_proba_pred = clf.predict_proba(X)

y_pred = np.where(y_proba_pred[:,1]<0.5,0,1)
AUROC = roc_auc_score(y, y_proba_pred[:,1])
print("AUROC: ", AUROC)

# calculate precision-recall curve
precision, recall, thresholds = precision_recall_curve(y, y_proba_pred[:,1])
AUPRC = auc(recall, precision)
print("AUPRC: ", AUPRC)

# calculate f1 score
f1 = f1_score(y, y_pred)
print("f1: ", f1)

# calculate accuracy score
accuracy = accuracy_score(y, y_pred)
print("accuracy: ", accuracy)

# calculate precision score
precision = precision_score(y, y_pred)
print("precision: ", precision)

# calculate recall score
recall = recall_score(y, y_pred)
print("recall: ", recall)


# using MHCflurry -PS
y = PS_benchmark_df['hit'].to_numpy()
y_pred = PS_benchmark_df['PS_y'].to_numpy()
y_proba_pred = PS_benchmark_df['PS_proba'].to_numpy()

AUROC = roc_auc_score(y, y_proba_pred)
print("AUROC: ", AUROC)

# calculate precision-recall curve
precision, recall, thresholds = precision_recall_curve(y, y_proba_pred)
AUPRC = auc(recall, precision)
print("AUPRC: ", AUPRC)

# calculate f1 score
f1 = f1_score(y, y_pred)
print("f1: ", f1)

# calculate accuracy score
accuracy = accuracy_score(y, y_pred)
print("accuracy: ", accuracy)

# calculate precision score
precision = precision_score(y, y_pred)
print("precision: ", precision)

# calculate recall score
recall = recall_score(y, y_pred)
print("recall: ", recall)
